+++ src/middlewares/index.js
@@ -1,0 +1,10 @@
+import authorized from './authorized';
+
+const middlewares = {
+  authorized,
+};
+
+export {
+  authorized,
+  middlewares as default,
+};