+++ src/models/user.js
@@ -1,9 +1,9 @@
 import mongoose from 'mongoose';
 
 const userSchema = new mongoose.Schema({
-  username: { type: String, required: true, unique: true },
-  password: { type: String, required: true },
-  api_token: String,
+  username: { type: String, unique: true },
+  password: String,
+  api_token: { type: String, unique: true },
   created_at: Date,
   updated_at: Date,
 });