+++ src/routes/index.js
@@ -1,10 +1,13 @@
 import login from './login';
+import user from './user';
 
 const routes = {
   login,
+  user,
 };
 
 export {
   login,
+  user,
   routes as default,
 };