+++ src/server.js
@@ -1,8 +1,9 @@
 import express from 'express';
 import passport from 'passport';
 import mongoose from 'mongoose';
-import { login as loginRoute } from './routes';
+import routes from './routes';
 import { toPromise } from './util';
+import init from './init';
 import './controllers';
 
 const mongoUrl = process.env.MONGO_URL || 'mongodb://db';
@@ -14,13 +15,16 @@ app.use(passport.initialize());
 app.use(passport.session());
 
 // Setup routing
-app.use('/api/login', loginRoute);
+Object.keys(routes).forEach(route => app.use(`/api/${route}`, routes[route]));
 
 // Connect mongoose
 mongoose.connect(mongoUrl);
 
 // Start the webserver
 toPromise(app, app.listen, port)
+  .then(
+    () => init()
+  )
   .then(
     () => console.info(`Server started listening on port ${port}!`),
     error => console.error(`Error occurred whilst starting the server: ${error}`)