+++ src/util/index.js
@@ -1,10 +1,13 @@
 import toPromise from './to-promise';
+import fileExists from './file-exists';
 
 const util = {
   toPromise,
+  fileExists,
 };
 
 export {
   toPromise,
+  fileExists,
   util as default,
 };